+++ src/project.rs
@@ -106,7 +106,7 @@ impl Project {
 
             let data = get_min_maxes(raw_data, nframes, width as usize);
             let (all_min, all_max, minmaxes) = data.clone();
-            let vscale = (height as f32 / 2.0) / f32::max(i16::abs(all_min) as f32, i16::abs(all_max) as f32);
+            let vscale = (height as f32 / 2.0) / f32::max(f32::abs(all_min), f32::abs(all_max));
 
             // UH-UH! Need to get rid of fg.clone()
             let mut wf_img = WaveformImg::new(width, height, vscale, fg.clone(), bg);
